Create ChallengeProject1
This was a challenge project I set for myself in my first month of the program at Keyin. I wanted to develop something that had the functionality of a program you might encounter in the real world, within the bounds of what I had learned in the course and the text book I had been recommended by my instructor.